refactor: merge client and server directories
