+++ .babelrc
@@ -1,4 +1,6 @@
 {
+  // Custom babel preset to make class decorators work with Next JS
+  // Taken from https://github.com/vercel/next.js/issues/4707#issuecomment-663980951
   "presets": ["next/babel"],
   "plugins": [
     "babel-plugin-transform-typescript-metadata",
