+++ .env.example
@@ -1,0 +1,10 @@
+NEXT_PUBLIC_FIREBASE_APIKEY=
+NEXT_PUBLIC_FIREBASE_MSG_SENDER_ID=
+NEXT_PUBLIC_FIREBASE_APP_ID=
+NEXT_PUBLIC_FIREBASE_MEASUREMENT_ID=
+
+FIREBASE_CLIENT_EMAIL=
+FIREBASE_PRIVATE_KEY=
+FIREBASE_PROJECT_ID=
+
+MONGODB_URL=
