--- client/.env.example
@@ -1,1 +1,0 @@
-NEXT_PUBLIC_API_URL=localhost:4000
