--- client/src/middleware/index.ts
@@ -1,2 +1,0 @@
-export { default as authMiddleware } from "./auth";
-export { default as withDatabase } from "./database";
