+++ package.json
@@ -1,18 +1,47 @@
 {
   "name": "aip_assessment_2",
+  "private": true,
   "scripts": {
-    "postinstall": "(cd client && npm install); (cd server && npm install);",
-    "prettier": "prettier . -w"
+    "prettier": "prettier . -w",
+    "dev": "next dev",
+    "dev:debug": "cross-env NODE_OPTIONS='--inspect' next dev",
+    "build": "next build",
+    "start": "next start"
   },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/LonJonn/AIP-Assessment-2.git"
   },
+  "dependencies": {
+    "@chakra-ui/core": "^1.0.0-rc.4",
+    "@typegoose/typegoose": "^7.4.1",
+    "firebase": "^7.21.1",
+    "firebase-admin": "^9.2.0",
+    "mongoose": "^5.10.7",
+    "next": "9.5.3",
+    "next-connect": "^0.9.0",
+    "react": "16.13.1",
+    "react-dom": "16.13.1",
+    "react-firebase-hooks": "^2.2.0",
+    "react-hook-form": "^6.8.6",
+    "swr": "^0.3.5",
+    "winston": "^3.3.3",
+    "yup": "^0.29.3"
+  },
   "devDependencies": {
-    "@commitlint/cli": "^9.1.2",
-    "@commitlint/config-conventional": "^9.1.2",
+    "@babel/plugin-proposal-decorators": "^7.10.5",
+    "@commitlint/cli": "^11.0.0",
+    "@commitlint/config-conventional": "^11.0.0",
+    "@types/mongoose": "^5.7.36",
+    "@types/node": "^14.11.2",
+    "@types/react": "^16.9.49",
+    "@types/yup": "^0.29.7",
+    "babel-plugin-parameter-decorator": "^1.0.16",
+    "babel-plugin-transform-typescript-metadata": "^0.3.0",
+    "cross-env": "^7.0.2",
     "husky": "^4.3.0",
-    "prettier": "^2.1.1"
+    "prettier": "^2.1.2",
+    "typescript": "^4.0.3"
   },
   "husky": {
     "hooks": {
