--- server/.env.example
@@ -1,3 +1,0 @@
-NODE_ENV=development
-MONGODB_URL=
-MONGODB_DB=
