--- server/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "aip_assessment_2_server",
-  "main": "app.js",
-  "scripts": {
-    "dev": "nodemon src/app.ts",
-    "dev:debug": "nodemon src/app.ts --inspect"
-  },
-  "dependencies": {
-    "@typegoose/typegoose": "^7.4.0",
-    "cors": "^2.8.5",
-    "dotenv": "^8.2.0",
-    "express": "^5.0.0-alpha.8",
-    "express-validator": "^6.6.1",
-    "firebase-admin": "^9.2.0",
-    "mongoose": "^5.10.5",
-    "winston": "^3.3.3"
-  },
-  "devDependencies": {
-    "@types/cors": "^2.8.7",
-    "@types/express": "^4.17.8",
-    "@types/mongoose": "^5.7.36",
-    "@types/node": "^14.11.1",
-    "nodemon": "^2.0.4",
-    "ts-node": "^9.0.0",
-    "typescript": "^4.0.2"
-  }
-}
