--- server/src/env.d.ts
@@ -1,7 +1,0 @@
-declare namespace NodeJS {
-  export interface ProcessEnv {
-    NODE_ENV: "production" | "development";
-    MONGODB_URL: string;
-    MONGODB_DB: string;
-  }
-}
