--- server/src/middleware/index.ts
@@ -1,1 +1,0 @@
-export { default as authMiddleware } from "./auth";
