--- server/src/utils/index.ts
@@ -1,2 +1,0 @@
-export { default as logger } from "./logger";
-export { default as errorHandler } from "./errorHandler";
