+++ src/lib/firebase/client.ts
