+++ src/lib/middleware/authMiddleware.ts
