+++ src/lib/middleware/index.ts
@@ -1,0 +1,2 @@
+export { default as authMiddleware } from "./authMiddleware";
+export { default as withDatabase } from "./database";
