+++ src/utils/logger.ts
@@ -1,6 +1,6 @@
 import winston from "winston";
 
-// Setup logging
+// Setup logging and output files
 const logger = winston.createLogger({
   level: "info",
   format: winston.format.combine(
@@ -13,11 +13,13 @@ const logger = winston.createLogger({
   ],
 });
 
+// Custom logger output for the console that looks 👌
 const customFormat = winston.format.printf(({ level, message, timestamp }) => {
   const formattedTime = timestamp.replace(/T/, " ").replace(/\..+/, "");
   return `[${formattedTime}] ${level}: ${message}`;
 });
 
+// I like colours
 winston.addColors({
   error: "red",
   info: "cyan",
